% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.calc.sdg.conc.R
\name{def.calc.sdg.conc}
\alias{def.calc.sdg.conc}
\title{Dissolved Gas Concentration Calculations}
\usage{
def.calc.sdg.conc(inputFile, volGas = "gasVolume", volH2O = "waterVolume",
  baro = "barometricPressure", waterTemp = "waterTemp",
  headspaceTemp = "headspaceTemp", eqCO2 = "concentrationCO2Gas",
  sourceCO2 = "concentrationCO2Air", eqCH4 = "concentrationCH4Gas",
  sourceCH4 = "concentrationCH4Air", eqN2O = "concentrationN2OGas",
  sourceN2O = "concentrationN2OAir")
}
\arguments{
\item{inputFile}{Name of the data fram containing the information needed to calculate the 
dissolved gas concentrations. If the headers are named: "gasVolume", "waterVolume", 
"barometricPressure", "waterTemp", "concentrationCO2Gas", "concentrationCO2Air", 
"concentrationCH4Gas", "concentrationCH4Air", "concentrationN2OGas", "concentrationN2OAir", 
respectively, no other inputs are required. Otherwise, the names of the columns need to be 
input for the function to work.}

\item{volGas}{Column name containing the data for volume of air equilibrated with 
water (mL) [string]}

\item{volH2O}{Column name containing the data for volume of water equilibrated with 
air (mL) [string]}

\item{baro}{Column name containing the data for barometric pressure at the time of 
equilibration (kPa) [string]}

\item{waterTemp}{Column name containing the data for temperature of the waterbody when 
sampled (celsius) [string]}

\item{headspaceTemp}{Column name containing the data for temperature of the water sample 
during the headspace equilibration (celsius) [string]}

\item{eqCO2}{Column name containing the data for concentration of carbon dioxide in the 
equilibrated gas (ppmv) [string]}

\item{sourceCO2}{Column name containing the data for concentration of carbon dioxide in 
headspace source gas (ppmv) [string]}

\item{eqCH4}{Column name containing the data for concentration of methane in the 
equilibrated gas (ppmv) [string]}

\item{sourceCH4}{Column name containing the data for concentration methane in headspace 
source gas (ppmv) [string]}

\item{eqN2O}{Column name containing the data for concentration of nitrous oxide in the 
equilibrated gas (ppmv) [string]}

\item{sourceN2O}{Column name containing the data for concentration of nitrous oxide in 
headspace source gas (ppmv) [string]}
}
\value{
This function returns dissolved CO2, CH4, and N2O concentrations in surface water 
[M] based on headspace equilibration data. Function also returns dissolved 100% saturation 
concentrations  [M] of CO2, CH4, and N2O in surface waters. Outputs are appended as 
additional columns to the input data frame
}
\description{
This function calculates dissolved CO2, CH4, and N2O concentrations from 
water samples based on inputs of equilibration conditions and reference and equilibrated 
air CO2, CH4, and N2O concentrations. If samples were equilibrated with a pure gas that 
contains no CO2, CH4, or N2O, the concentrations for the reference air 
("concentrationCO2Air", "concentrationCH4Air", "concentrationN2OAir") for those gases 
should be set to 0.
}
\examples{
#where the data frame "sdgFormatted" is already read in
#sdgDataPlusVals <- def.calc.sdg.conc(inputFile = sdgFormatted)
#where the data is read in from a file in the working directory (also works with a full path)
#sdgDataPlusVals <- def.calc.sdg.conc(inputFile = 
#system.file("extdata", "sdgTestData.csv", package = "neonDissGas"))
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
def.format.sdg.R for formatting dissolved gas data downloaded from NEON
}
\author{
Kaelin M. Cawley \email{kcawley@battelleecology.org} \cr
}
\keyword{CH4,}
\keyword{CO2,}
\keyword{N2O,}
\keyword{aquatic,}
\keyword{carbon}
\keyword{dioxide,}
\keyword{dissolved}
\keyword{gases,}
\keyword{lakes,}
\keyword{methane,}
\keyword{nitrous}
\keyword{oxide,}
\keyword{rivers}
\keyword{streams,}
\keyword{surface}
\keyword{water,}
