% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.calc.sdg.R
\name{def.calc.sdg}
\alias{def.calc.sdg}
\title{Dissolved Gas Concentration Calculations}
\usage{
def.calc.sdg(inputFile, volGas = inputFile$gasVolume,
  volH2O = inputFile$waterVolume, baro = inputFile$barometricPressure,
  waterTemp = inputFile$waterTemp, eqCO2 = inputFile$concentrationCO2Gas,
  airCO2 = inputFile$concentrationCO2Air,
  eqCH4 = inputFile$concentrationCH4Gas,
  airCH4 = inputFile$concentrationCH4Air,
  eqN2O = inputFile$concentrationN2OGas,
  airN2O = inputFile$concentrationN2OAir)
}
\arguments{
\item{inputFile}{Name of the data fram containing the information needed to calculate the dissolved gas concentrations. If the headers are named: "gasVolume", "waterVolume", "barometricPressure", "waterTemp", "concentrationCO2Gas", "concentrationCO2Air", "concentrationCH4Gas", "concentrationCH4Air", "concentrationN2OGas", "concentrationN2OAir", respectively, no other inputs are required. Otherwise, the names of the columns need to be input for the function to work.}

\item{volGas}{Volume of air equilibrated with water [mL]}

\item{volH2O}{Volume of water equilibrated with air [mL]}

\item{baro}{Barometric pressure at the time of equilibration [kPa]}

\item{waterTemp}{Temperature of the water at the time of equilibration [celsius]}

\item{eqCO2}{Concentration of carbon dioxide in the equilibrated gas [ppmv]}

\item{airCO2}{Concentration of carbon dioxide in the reference air [ppmv]}

\item{eqCH4}{Concentration of methane in the equilibrated gas [ppmv]}

\item{airCH4}{Concentration of methane in the reference air [ppmv]}

\item{eqN2O}{Concentration of nitrous oxide in the equilibrated gas [ppmv]}

\item{airN2O}{Concentration of nitrous oxide in the reference air [ppmv]}
}
\value{
This function returns dissolved CO2, CH4, and N2O concentrations in surface water [M] appended as additional columns to the input data frame
}
\description{
This function calculates dissolved CO2, CH4, and N2O concentrations from water samples based on inputs of equilibration conditions and reference and equilibrated air CO2, CH4, and N2O concentrations. If samples were equilibrated with a pure gas that contains no CO2, CH4, or N2O, the concentrations for the reference air ("concentrationCO2Air", "concentrationCH4Air", "concentrationN2OAir") for those gases should be set to 0.
}
\examples{
#where the data frame "sdgFormatted" is already read in
sdgDataPlusVals <- def.calc.sdg(inputFile = sdgFormatted)
#where the data is read in from a file in the working directory (also works with a full path)
sdgDataPlusVals <- def.calc.sdg(inputFile = 
system.file("extdata", "sdgTestData.csv", package = "neonDissGas"))
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
def.format.sdg.R for formatting dissolved gas data downloaded from NEON
}
\author{
Kaelin M. Cawley \email{kcawley@battelleecology.org} \cr
}
\keyword{CH4,}
\keyword{CO2,}
\keyword{N2O,}
\keyword{aquatic,}
\keyword{carbon}
\keyword{dioxide,}
\keyword{dissolved}
\keyword{gases,}
\keyword{lakes,}
\keyword{methane,}
\keyword{nitrous}
\keyword{oxide,}
\keyword{rivers}
\keyword{streams,}
\keyword{surface}
\keyword{water,}
