% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.calc.sdg.sat.R
\name{def.calc.sdg.sat}
\alias{def.calc.sdg.sat}
\title{Dissolved Gas Percent Saturation Calculations}
\usage{
def.calc.sdg.sat(inputFile, baro = "barometricPressure",
  waterTemp = "waterTemp", headspaceTemp = "headspaceTemp",
  concCO2 = "dissolvedCO2", sourceCO2 = "concentrationCO2Air",
  concCH4 = "dissolvedCH4", sourceCH4 = "concentrationCH4Air",
  concN2O = "dissolvedN2O", sourceN2O = "concentrationN2OAir")
}
\arguments{
\item{inputFile}{Name of the data fram containing the information needed to calculate 
the dissolved gas percent saturation. If the headers are named: barometricPressure", 
"waterTemp", "dissolvedCO2", "concentrationCO2Air", "dissolvedCH4", "concentrationCH4Air", 
"dissolvedN2O", "concentrationN2OAir", respectively, no other inputs are required. Otherwise, 
the names of the columns need to be input for the function to work.}

\item{baro}{Column name containing the data for barometric pressure at the time of 
equilibration (kPa) [string]}

\item{waterTemp}{Column name containing the data for temperature of the waterbody when 
sampled (celsius) [string]}

\item{headspaceTemp}{Column name containing the data for temperature of the water sample 
during the headspace equilibration (celsius) [string]}

\item{concCO2}{Column name containing the data for concentration of carbon dioxide in the 
water (M) [string]}

\item{sourceCO2}{Column name containing the data for concentration of carbon dioxide in 
headspace source gas (ppmv) [string]}

\item{concCH4}{Column name containing the data for concentration of methane in the 
water (M) [string]}

\item{sourceCH4}{Column name containing the data for concentration methane in headspace 
source gas (ppmv) [string]}

\item{concN2O}{Column name containing the data for concentration of nitrous oxide in the 
water (M) [string]}

\item{sourceN2O}{Column name containing the data for concentration of nitrous oxide in 
headspace source gas (ppmv) [string]}
}
\value{
This function returns dissolved CO2, CH4, and N2O concentrations in surface water [M] based on headspace equilibration data.
  Function also returns dissolved 100% saturation concentrations  [M] of CO2, CH4, and N2O in surface waters.
  Outputs are appended as additional columns to the input data frame
}
\description{
This function calculates dissolved CO2, CH4, and N2O percent saturation from 
water dissolved gas concentration (molar), water temperature (celsius), barometric pressure 
(kPa), and reference air gas concentrations (ppmv).
}
\examples{
#where the data frame "sdgFormatted" is already read in
#sdgDataPlusSat <- def.calc.sdg.sat(inputFile = sdgFormatted)
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
def.format.sdg.R for formatting dissolved gas data downloaded from NEON
}
\author{
Kaelin M. Cawley \email{kcawley@battelleecology.org} \cr
}
\keyword{CH4,}
\keyword{CO2,}
\keyword{N2O,}
\keyword{aquatic,}
\keyword{carbon}
\keyword{dioxide,}
\keyword{dissolved}
\keyword{gases,}
\keyword{lakes,}
\keyword{methane,}
\keyword{nitrous}
\keyword{oxide,}
\keyword{rivers}
\keyword{streams,}
\keyword{surface}
\keyword{water,}
