% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.format.sdg.R
\name{def.format.sdg}
\alias{def.format.sdg}
\title{Dissolved Gas Data Formatting}
\usage{
def.format.sdg(externalLabData = NA, fieldDataProc = NA, fieldSuperParent = NA)
}
\arguments{
\item{externalLabData}{Data frame containing external lab data for NEON}

\item{fieldDataProc}{Data frame containing field processing data for NEON}

\item{fieldSuperParent}{Data frame containing field collection data for NEON}
}
\value{
This function returns one data frame formatted for use with def.calc.sdg.R to actually calculate the concentration of the gases in the surface water sample
}
\description{
This function reads in data from the NEON Dissolved Gas data product to calculate dissolved gas concentrations in surface water. For the best results download the expanded dissolved gas package. No need to unzip the downloaded files, just place them all in the smae directory.
}
\examples{
\dontrun{TBD}
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
def.calc.sdg.conc.R and def.calc.sdg.sat.R for calculating dissolved gas 
concentrations and percent saturation, respectively
CO2BelowDetection',
CH4BelowDetection',
N2OBelowDetection',
}
\author{
Kaelin M. Cawley \email{kcawley@battelleecology.org} \cr
}
\keyword{CH4,}
\keyword{CO2,}
\keyword{N2O,}
\keyword{aquatic,}
\keyword{carbon}
\keyword{dioxide,}
\keyword{dissolved}
\keyword{gases,}
\keyword{lakes,}
\keyword{methane,}
\keyword{nitrous}
\keyword{oxide,}
\keyword{rivers}
\keyword{streams,}
\keyword{surface}
\keyword{water,}
