% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.format.sdg.R
\name{def.format.sdg}
\alias{def.format.sdg}
\title{Dissolved Gas Data Formatting}
\usage{
def.format.sdg(dataDir = paste0(getwd(),
  "/NEON_dissolved-gases-surfacewater.zip"))
}
\arguments{
\item{dataDir}{User identifies the directory that contains the zipped data}
}
\value{
This function returns one data frame formatted for use with def.calc.sdg.R to actually calculate the concentration of the gases in the surface water sample
}
\description{
This function reads in data from the NEON Dissolved Gas data product to calculate dissolved gas concentrations in surface water. For the best results download the expanded dissolved gas package. No need to unzip the downloaded files, just place them all in the smae directory.
}
\examples{
#where the data .zip file is in the working directory and has the default name, 
#sdgFormatted <- def.format.sdg()
#where the data.zip file is in the downloads folder and has default name, 
#sdgFormatted <- 
#def.format.sdg(dataDir = path.expand("~/Downloads/NEON_dissolved-gases-surfacewater.zip"))
#where the data.zip file is in the downloads folder and has a specified name,
#sdgFormatted <- def.format.sdg(dataDir = path.expand("~/Downloads/non-standard-name.zip"))
#Using the example data in this package
dataDirectory <- paste(path.package("neonDissGas"),"inst\\\\extdata", sep = "\\\\")
sdgFormatted <- def.format.sdg(dataDir = dataDirectory)
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
def.values.sdg.R and def.calc.sdg.R for calculating dissolved gas concentrations
CO2BelowDetection',
CH4BelowDetection',
N2OBelowDetection',
}
\author{
Kaelin M. Cawley \email{kcawley@battelleecology.org} \cr
}
\keyword{CH4,}
\keyword{CO2,}
\keyword{N2O,}
\keyword{aquatic,}
\keyword{carbon}
\keyword{dioxide,}
\keyword{dissolved}
\keyword{gases,}
\keyword{lakes,}
\keyword{methane,}
\keyword{nitrous}
\keyword{oxide,}
\keyword{rivers}
\keyword{streams,}
\keyword{surface}
\keyword{water,}
